pyproject.toml
[tool.poetry]
name = "projex"
version = "0.0.1"
description = ""
authors = ["actusnileh <actusnileh@gmail.com>"]
license = "GNU"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
pydantic-settings = "^2.6.1"
debugpy = "^1.8.8"
uvicorn = "^0.32.0"
punq = "^0.7.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
motor = "^3.6.0"
faker = "^30.8.2"
httpx = "^0.27.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
multi_line_output=3
include_trailing_comma=true
line_length=120
lines_after_imports=2
lines_between_sections=1
lines_between_types=0
force_alphabetical_sort_within_sections=true
ensure_newline_before_comments=true
combine_as_imports=true
combine_star=true
skip_gitignore=true
balanced_wrapping=true
force_grid_wrap=2
src_paths=["src", "tests"]
skip_glob=[
    ".git",
    ".idea",
    ".vscode",
    "venv",
    ".venv",
]
known_fastapi=["fastapi"]
known_first_party=["src"]
sections=[
    "FUTURE",
    "STDLIB",
    "FASTAPI",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]